refactor(blockdigit): split blockdigit into core / manipulators / iostream (#1334)
blockdigit.hpp carried <iostream>, <sstream> and <iomanip> for three functions at the
bottom of the file -- operator<<, type_tag and to_binary -- while the 620 lines above them
are arithmetic. It is the storage type behind blockoctal, blockdecimal, blockhexadecimal,
positional and rational, so all five inherited those headers.

Now matching blockbinary's existing shape:

    blockdigit.hpp        arithmetic, 0 I/O-family
    manipulators.hpp      type_tag / to_binary
    iostream.hpp          operator<<

The to_string() member stays in the core: it builds a std::string character by character
and never needed a stream. The divide-by-zero diagnostic moved to fprintf(stderr, ...),
the Phase 0 idiom, and the friend declaration needs only <iosfwd>.

Consumers updated to include what they now use: the blockdigit api test, and
positional_impl.hpp, whose own operator<< streams its Storage through blockdigit's.

positional/manipulators.hpp had NO includes at all while using std::string,
std::stringstream and to_binary(blockdigit) -- it compiled only behind a prior include.
Same defect class as #1422, fixed here because this change is what exposed it.

Verified: every api.cpp in the tree, 55 pass 0 fail, on gcc and clang.
